Self-contained FFNN: Sequential, MultiHead, and a training loop
Rebuilds the neural network stack inside neural_networks/ without the
dolcestat.core refactor. The package now owns everything it needs; the
only import from the rest of the library is DolceSet. optimization/,
linear_models/, metrics/, neighbors/, trees/ and preprocessing/ are
byte-identical to 96e03ef.

New:
- optimizers.py: an Optimizer ABC declaring step(parameters), and a
  deliberately minimal GradientDescent (learning rate only, no momentum,
  no flavors). This is what makes Sequential.fit work at all -- it has
  always called optimizer.step(), which nothing implemented.
- MultiHead: a shared trunk feeding N heads, summing their gradients
  back into the trunk. fit/predict move up to NeuralNetwork so both
  containers inherit one loop.
- MeanSquaredError, GaussianNegativeLogLikelihood, Softplus.
- fit(track_history=True) records the per-epoch loss, read back with
  get_loss(iteration=None) -- the same accessor BaseOptimizer exposes.

Fixed:
- Sigmoid saturation: the clip now lives in Sigmoid.forward and the
  cached array is the one returned, so sigmoid+BCE cancel exactly. A
  saturated unit previously got a zero gradient instead of (A - y)/n,
  which silently stopped XOR from learning.
- Targets are reshaped to a column once, in the loop, rather than inside
  each loss; losses now assert matching shapes instead.
- Parameters.grad is float-typed regardless of the value's dtype.
- Mutable defaults in DenseLayer and fit resolved to None sentinels.

Verified by finite-difference gradient checks on Sequential and
MultiHead (worst relative error ~1e-9), a saturation regression test,
and notebooks 00-09 all executing clean.

+++ src/dolcestat/neural_networks/parameters.py
@@ -15,7 +15,10 @@ class Parameters:
 
     def __init__(self, value):
         self.value = value
-        self.grad = np.zeros_like(value)
+        # dtype=float explicitly: np.zeros_like would otherwise inherit an
+        # integer dtype from an integer-valued parameter, and the first
+        # ``grad += <float array>`` in a backward pass would raise.
+        self.grad = np.zeros_like(value, dtype=float)
 
     def zero_grad(self):
         "Zeroes the stored gradient. Called at the start of each epoch."
